Check for Cell in railtie
