+++ lib/trailblazer/rails.rb
@@ -8,6 +8,7 @@ end
 require "reform/rails"
 
 require "trailblazer/rails/controller"
+require "trailblazer/rails/controller/cell"
 require "trailblazer/rails/form"
 require "trailblazer/rails/railtie"
 require "trailblazer"
